+++ src/TSVQ.h
@@ -99,7 +99,7 @@ public:
 private:
     double RMSE() {
         double RMSE = sumSquaredError(NULL, _root);
-        int size = getObjCount();
+        uint64_t size = getObjCount();
         RMSE /= size;
         RMSE = sqrt(RMSE);
         return RMSE;
